% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{compactify_edges}
\alias{compactify_edges}
\title{Construct indices matrices}
\usage{
compactify_edges(w, n, method = "ama")
}
\arguments{
\item{w}{weights vector}

\item{n}{number of points to cluster}

\item{method}{'ama' or 'admm'}
}
\description{
\code{compactify_edges} constructs M1, M2, and ix index matrices. Note that storage conventions are different for ama and admm.
}

