% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{create_clustering_problem}
\alias{create_clustering_problem}
\title{Create a random clustering problem}
\usage{
create_clustering_problem(p, n, seed = 12345, nnn = 3, method = "ama")
}
\arguments{
\item{p}{Dimension of space of points to be clustered}

\item{n}{Number of points}

\item{seed}{Random number seed}

\item{nnn}{Number of nearest neighbors}

\item{method}{'ama' or 'admm'}
}
\description{
\code{create_clustering_problem} makes a random clustering problem for testing purposes.
}
\examples{
p <- 10
n <- 20
seed <- 12345
rnd_problem_admm <- create_clustering_problem(p,n,seed)
}

